move tap dance logic to separate file
